Remove tests based on the old functionality, and add new better tests.
