+++ libs/core/kiln_ai/adapters/model_adapters/test_litellm_adapter.py
@@ -682,342 +682,3 @@ async def test_build_completion_kwargs_raises_error_with_tools_conflict(
         else:
             # should not raise an error
             await adapter.build_completion_kwargs(mock_provider, messages, None)
-
-
-@pytest.mark.asyncio
-async def test_handle_tool_calls_loop_no_tool_calls(config, mock_task):
-    """Test _handle_tool_calls_loop with no tool calls in response"""
-    adapter = LiteLlmAdapter(config=config, kiln_task=mock_task)
-    initial_messages = [{"role": "user", "content": "Hello"}]
-
-    # Mock response with no tool calls
-    mock_response = Mock()
-    mock_choice = Mock()
-    mock_message = Mock()
-    mock_message.content = "Hello response"
-    mock_message.tool_calls = None
-    mock_choice.message = mock_message
-    mock_response.choices = [mock_choice]
-
-    with patch.object(
-        adapter,
-        "acompletion_checking_response",
-        return_value=(mock_response, mock_choice),
-    ):
-        content, messages = await adapter._handle_tool_calls_loop(
-            Mock(), initial_messages, None
-        )
-
-    assert content == "Hello response"
-    assert len(messages) == 2
-    assert messages[0] == {"role": "user", "content": "Hello"}
-    assert messages[1] == {
-        "role": "assistant",
-        "content": "Hello response",
-        "tool_calls": None,
-    }
-
-
-@pytest.mark.asyncio
-async def test_handle_tool_calls_loop_with_task_response_tool(config, mock_task):
-    """Test _handle_tool_calls_loop with task_response tool call"""
-    adapter = LiteLlmAdapter(config=config, kiln_task=mock_task)
-    initial_messages = [{"role": "user", "content": "Hello"}]
-
-    # Mock response with task_response tool call
-    mock_response = Mock()
-    mock_choice = Mock()
-    mock_message = Mock()
-    mock_message.content = None
-    mock_choice.message = mock_message
-    mock_response.choices = [mock_choice]
-
-    # Mock tool call
-    mock_tool_call = Mock()
-    mock_tool_call.function.name = "task_response"
-    mock_tool_call.function.arguments = '{"result": "final output"}'
-    mock_message.tool_calls = [mock_tool_call]
-
-    with patch.object(
-        adapter,
-        "acompletion_checking_response",
-        return_value=(mock_response, mock_choice),
-    ):
-        content, messages = await adapter._run_model_turn(
-            Mock(), initial_messages, None, False
-        )
-
-    assert content == '{"result": "final output"}'
-    assert len(messages) == 2
-    assert messages[0] == {"role": "user", "content": "Hello"}
-    assert messages[1] == {
-        "role": "assistant",
-        "content": None,
-        "tool_calls": [mock_tool_call],
-    }
-
-
-@pytest.mark.asyncio
-async def test_handle_tool_calls_loop_with_normal_tool_calls(
-    config, mock_task, mock_math_tools
-):
-    """Test _handle_tool_calls_loop with normal tool calls"""
-    adapter = LiteLlmAdapter(config=config, kiln_task=mock_task)
-    initial_messages = [{"role": "user", "content": "Calculate 2 + 3"}]
-
-    # Mock response with normal tool call
-    mock_response = Mock()
-    mock_choice = Mock()
-    mock_message = Mock()
-    mock_message.content = None
-    mock_choice.message = mock_message
-    mock_response.choices = [mock_choice]
-
-    # Mock tool call
-    mock_tool_call = Mock()
-    mock_tool_call.function.name = "add"
-    mock_tool_call.function.arguments = '{"a": 2, "b": 3}'
-    mock_tool_call.id = "tool_call_1"
-    mock_message.tool_calls = [mock_tool_call]
-
-    # Mock second response with final content
-    mock_response2 = Mock()
-    mock_choice2 = Mock()
-    mock_message2 = Mock()
-    mock_message2.content = "The result is 5"
-    mock_message2.tool_calls = None
-    mock_choice2.message = mock_message2
-    mock_response2.choices = [mock_choice2]
-
-    with (
-        patch.object(adapter, "cached_available_tools", return_value=mock_math_tools),
-        patch.object(
-            adapter,
-            "acompletion_checking_response",
-            side_effect=[(mock_response, mock_choice), (mock_response2, mock_choice2)],
-        ),
-    ):
-        content, messages = await adapter._handle_tool_calls_loop(
-            Mock(), initial_messages, None
-        )
-
-    assert content == "The result is 5"
-    # The method should add:
-    # 1. The initial message
-    # 2. The assistant's response with tool calls
-    # 3. The tool call result
-    # 4. The final assistant response
-    assert len(messages) == 4
-    assert messages[0] == {"role": "user", "content": "Calculate 2 + 3"}
-    assert messages[1] == {
-        "role": "assistant",
-        "content": None,
-        "tool_calls": [mock_tool_call],
-    }
-    assert messages[2] == {
-        "role": "tool",
-        "tool_call_id": "tool_call_1",
-        "name": "add",
-        "content": "5",
-    }
-    assert messages[3] == {
-        "role": "assistant",
-        "content": "The result is 5",
-        "tool_calls": None,
-    }
-
-
-@pytest.mark.asyncio
-async def test_handle_tool_calls_loop_multiple_turns(
-    config, mock_task, mock_math_tools
-):
-    """Test _handle_tool_calls_loop with multiple turns of tool calls"""
-    adapter = LiteLlmAdapter(config=config, kiln_task=mock_task)
-    initial_messages = [{"role": "user", "content": "Calculate (2 + 3) * 4"}]
-
-    # Mock first response with add tool call
-    mock_response1 = Mock()
-    mock_choice1 = Mock()
-    mock_message1 = Mock()
-    mock_message1.content = None
-    mock_choice1.message = mock_message1
-    mock_response1.choices = [mock_choice1]
-
-    mock_tool_call1 = Mock()
-    mock_tool_call1.function.name = "add"
-    mock_tool_call1.function.arguments = '{"a": 2, "b": 3}'
-    mock_tool_call1.id = "tool_call_1"
-    mock_message1.tool_calls = [mock_tool_call1]
-
-    # Mock second response with multiply tool call
-    mock_response2 = Mock()
-    mock_choice2 = Mock()
-    mock_message2 = Mock()
-    mock_message2.content = None
-    mock_choice2.message = mock_message2
-    mock_response2.choices = [mock_choice2]
-
-    mock_tool_call2 = Mock()
-    mock_tool_call2.function.name = "multiply"
-    mock_tool_call2.function.arguments = '{"a": 5, "b": 4}'
-    mock_tool_call2.id = "tool_call_2"
-    mock_message2.tool_calls = [mock_tool_call2]
-
-    # Mock third response with final content
-    mock_response3 = Mock()
-    mock_choice3 = Mock()
-    mock_message3 = Mock()
-    mock_message3.content = "The result is 20"
-    mock_message3.tool_calls = None
-    mock_choice3.message = mock_message3
-    mock_response3.choices = [mock_choice3]
-
-    with (
-        patch.object(adapter, "cached_available_tools", return_value=mock_math_tools),
-        patch.object(
-            adapter,
-            "acompletion_checking_response",
-            side_effect=[
-                (mock_response1, mock_choice1),
-                (mock_response2, mock_choice2),
-                (mock_response3, mock_choice3),
-            ],
-        ),
-    ):
-        content, messages = await adapter._handle_tool_calls_loop(
-            Mock(), initial_messages, None
-        )
-
-    assert content == "The result is 20"
-    # The method should add:
-    # 1. The initial message
-    # 2. The assistant's response with first tool call
-    # 3. The first tool call result
-    # 4. The assistant's response with second tool call
-    # 5. The second tool call result
-    # 6. The final assistant response
-    assert len(messages) == 6
-    assert messages[0] == {"role": "user", "content": "Calculate (2 + 3) * 4"}
-    assert messages[1] == {
-        "role": "assistant",
-        "content": None,
-        "tool_calls": [mock_tool_call1],
-    }
-    assert messages[2] == {
-        "role": "tool",
-        "tool_call_id": "tool_call_1",
-        "name": "add",
-        "content": "5",
-    }
-    assert messages[3] == {
-        "role": "assistant",
-        "content": None,
-        "tool_calls": [mock_tool_call2],
-    }
-    assert messages[4] == {
-        "role": "tool",
-        "tool_call_id": "tool_call_2",
-        "name": "multiply",
-        "content": "20",
-    }
-    assert messages[5] == {
-        "role": "assistant",
-        "content": "The result is 20",
-        "tool_calls": None,
-    }
-
-
-@pytest.mark.asyncio
-async def test_run_method_updates_chat_history(config, mock_task):
-    """Test _run method properly updates chat history"""
-    adapter = LiteLlmAdapter(config=config, kiln_task=mock_task)
-
-    # Mock chat formatter
-    mock_chat_formatter = Mock()
-    mock_turn = Mock()
-    mock_turn.messages = [
-        Mock(role="user", content="Hello"),
-        Mock(role="system", content="You are a helpful assistant"),
-    ]
-    mock_turn.final_call = True
-    # Make it return None on the second call to break the loop
-    mock_chat_formatter.next_turn.side_effect = [mock_turn, None]
-    mock_chat_formatter.intermediate_outputs.return_value = {}
-
-    # Mock response
-    mock_response = Mock()
-    mock_choice = Mock()
-    mock_message = Mock()
-    mock_message.content = "Hello response"
-    mock_message.reasoning_content = None
-    mock_message.tool_calls = None
-    mock_choice.message = mock_message
-    mock_response.choices = [mock_choice]
-
-    with (
-        patch.object(adapter, "build_chat_formatter", return_value=mock_chat_formatter),
-        patch.object(adapter, "litellm_tools", return_value=[]),
-        patch.object(adapter, "build_completion_kwargs", return_value={}),
-        patch.object(
-            adapter,
-            "acompletion_checking_response",
-            return_value=(mock_response, mock_choice),
-        ),
-    ):
-        await adapter._run("Hello")
-
-    # Chat history should have been updated
-    assert len(adapter._chat_history) == 3
-    assert adapter._chat_history[0] == {"role": "user", "content": "Hello"}
-    assert adapter._chat_history[1] == {
-        "role": "system",
-        "content": "You are a helpful assistant",
-    }
-    assert adapter._chat_history[2] == {
-        "role": "assistant",
-        "content": "Hello response",
-        "tool_calls": None,
-    }
-
-
-@pytest.mark.asyncio
-async def test_run_method_handles_tool_calls(config, mock_task, mock_math_tools):
-    """Test _run method properly handles tool calls when they are available"""
-    adapter = LiteLlmAdapter(config=config, kiln_task=mock_task)
-
-    # Mock chat formatter
-    mock_chat_formatter = Mock()
-    mock_turn = Mock()
-    mock_turn.messages = [Mock(role="user", content="Calculate 2 + 3")]
-    mock_turn.final_call = True
-    mock_chat_formatter.next_turn.return_value = mock_turn
-    mock_chat_formatter.intermediate_outputs.return_value = {}
-
-    # Mock response with tool call
-    mock_response = Mock()
-    mock_choice = Mock()
-    mock_message = Mock()
-    mock_message.content = "The result is 5"
-    mock_choice.message = mock_message
-    mock_response.choices = [mock_choice]
-
-    with (
-        patch.object(adapter, "build_chat_formatter", return_value=mock_chat_formatter),
-        patch.object(adapter, "litellm_tools", return_value=[Mock()]),
-        patch.object(
-            adapter,
-            "_handle_tool_calls_loop",
-            return_value=(
-                "The result is 5",
-                [
-                    {"role": "user", "content": "Calculate 2 + 3"},
-                    {"role": "assistant", "content": None, "tool_calls": [Mock()]},
-                ],
-            ),
-        ),
-    ):
-        output, usage = await adapter._run("Calculate 2 + 3")
-
-    assert output.output == "The result is 5"
-    # Chat history should have been updated
-    assert len(adapter.chat_history) == 2
